requirements.txt
spellchecker
gradio_client
streamlit
streamlit-chat
transformers
opencv-python
numpy
Pillow
pytesseract
pandas
PyPDF2
python-docx